Extracted math utils from api.
