+++ tests/unit/all/common/test_string_utils.c
@@ -1,0 +1,271 @@
+// Test suite for workspace/all/common/utils/string_utils.c
+// Tests string matching and manipulation functions
+
+#include "../../../../workspace/all/common/utils/string_utils.h"
+#include "../../../support/unity/unity.h"
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+// Unity requires these
+void setUp(void) {}
+void tearDown(void) {}
+
+// ============================================================================
+// String Matching Tests
+// ============================================================================
+
+void test_prefixMatch_exact(void) {
+	TEST_ASSERT_TRUE(prefixMatch("hello", "hello"));
+}
+
+void test_prefixMatch_prefix(void) {
+	TEST_ASSERT_TRUE(prefixMatch("hel", "hello"));
+}
+
+void test_prefixMatch_no_match(void) {
+	TEST_ASSERT_FALSE(prefixMatch("world", "hello"));
+}
+
+void test_prefixMatch_case_insensitive(void) {
+	TEST_ASSERT_TRUE(prefixMatch("HeLLo", "hello"));
+}
+
+void test_prefixMatch_empty(void) {
+	TEST_ASSERT_TRUE(prefixMatch("", "anything"));
+}
+
+void test_suffixMatch_exact(void) {
+	TEST_ASSERT_TRUE(suffixMatch("hello", "hello"));
+}
+
+void test_suffixMatch_suffix(void) {
+	TEST_ASSERT_TRUE(suffixMatch("llo", "hello"));
+}
+
+void test_suffixMatch_no_match(void) {
+	TEST_ASSERT_FALSE(suffixMatch("world", "hello"));
+}
+
+void test_suffixMatch_case_insensitive(void) {
+	TEST_ASSERT_TRUE(suffixMatch("LLO", "hello"));
+}
+
+void test_suffixMatch_extension(void) {
+	TEST_ASSERT_TRUE(suffixMatch(".txt", "file.txt"));
+	TEST_ASSERT_TRUE(suffixMatch(".disabled", "test.disabled"));
+}
+
+void test_exactMatch_same(void) {
+	TEST_ASSERT_TRUE(exactMatch("hello", "hello"));
+}
+
+void test_exactMatch_different(void) {
+	TEST_ASSERT_FALSE(exactMatch("hello", "world"));
+}
+
+void test_exactMatch_case_sensitive(void) {
+	TEST_ASSERT_FALSE(exactMatch("hello", "Hello"));
+}
+
+void test_exactMatch_null_strings(void) {
+	TEST_ASSERT_FALSE(exactMatch(NULL, "hello"));
+	TEST_ASSERT_FALSE(exactMatch("hello", NULL));
+	TEST_ASSERT_FALSE(exactMatch(NULL, NULL));
+}
+
+void test_containsString_found(void) {
+	TEST_ASSERT_TRUE(containsString("hello world", "world"));
+	TEST_ASSERT_TRUE(containsString("hello world", "hello"));
+	TEST_ASSERT_TRUE(containsString("hello world", "o w"));
+}
+
+void test_containsString_not_found(void) {
+	TEST_ASSERT_FALSE(containsString("hello world", "xyz"));
+}
+
+void test_containsString_case_insensitive(void) {
+	TEST_ASSERT_TRUE(containsString("Hello World", "WORLD"));
+	TEST_ASSERT_TRUE(containsString("Hello World", "world"));
+}
+
+void test_hide_hidden_file(void) {
+	TEST_ASSERT_TRUE(hide(".hidden"));
+	TEST_ASSERT_TRUE(hide(".gitignore"));
+}
+
+void test_hide_disabled_file(void) {
+	TEST_ASSERT_TRUE(hide("test.disabled"));
+	TEST_ASSERT_TRUE(hide("something.disabled"));
+}
+
+void test_hide_map_txt(void) {
+	TEST_ASSERT_TRUE(hide("map.txt"));
+}
+
+void test_hide_normal_file(void) {
+	TEST_ASSERT_FALSE(hide("normal.txt"));
+	TEST_ASSERT_FALSE(hide("test.c"));
+}
+
+// ============================================================================
+// String Manipulation Tests
+// ============================================================================
+
+void test_normalizeNewline_windows(void) {
+	char line[] = "test\r\n";
+	normalizeNewline(line);
+	TEST_ASSERT_EQUAL_STRING("test\n", line);
+}
+
+void test_normalizeNewline_unix(void) {
+	char line[] = "test\n";
+	normalizeNewline(line);
+	TEST_ASSERT_EQUAL_STRING("test\n", line);
+}
+
+void test_normalizeNewline_no_newline(void) {
+	char line[] = "test";
+	normalizeNewline(line);
+	TEST_ASSERT_EQUAL_STRING("test", line);
+}
+
+void test_trimTrailingNewlines_single(void) {
+	char line[] = "test\n";
+	trimTrailingNewlines(line);
+	TEST_ASSERT_EQUAL_STRING("test", line);
+}
+
+void test_trimTrailingNewlines_multiple(void) {
+	char line[] = "test\n\n\n";
+	trimTrailingNewlines(line);
+	TEST_ASSERT_EQUAL_STRING("test", line);
+}
+
+void test_trimTrailingNewlines_none(void) {
+	char line[] = "test";
+	trimTrailingNewlines(line);
+	TEST_ASSERT_EQUAL_STRING("test", line);
+}
+
+void test_trimSortingMeta_with_number(void) {
+	char buffer[] = "001) Game Name";
+	char* str = buffer;
+	trimSortingMeta(&str);
+	TEST_ASSERT_EQUAL_STRING("Game Name", str);
+}
+
+void test_trimSortingMeta_no_number(void) {
+	char buffer[] = "Game Name";
+	char* str = buffer;
+	char* original = str;
+	trimSortingMeta(&str);
+	TEST_ASSERT_EQUAL_STRING("Game Name", str);
+	TEST_ASSERT_EQUAL_PTR(original, str);
+}
+
+void test_trimSortingMeta_with_space(void) {
+	char buffer[] = "42)   Game";
+	char* str = buffer;
+	trimSortingMeta(&str);
+	TEST_ASSERT_EQUAL_STRING("Game", str);
+}
+
+// ============================================================================
+// Text Line Splitting Tests
+// ============================================================================
+
+void test_splitTextLines_single_line(void) {
+	char str[] = "Hello World";
+	char* lines[MAX_TEXT_LINES];
+	int count = splitTextLines(str, lines, MAX_TEXT_LINES);
+	TEST_ASSERT_EQUAL(1, count);
+	TEST_ASSERT_EQUAL_STRING("Hello World", lines[0]);
+}
+
+void test_splitTextLines_multiple_lines(void) {
+	char str[] = "Line 1\nLine 2\nLine 3";
+	char* lines[MAX_TEXT_LINES];
+	int count = splitTextLines(str, lines, MAX_TEXT_LINES);
+	TEST_ASSERT_EQUAL(3, count);
+	TEST_ASSERT_EQUAL_PTR(str, lines[0]);
+	TEST_ASSERT_EQUAL_STRING("Line 2\nLine 3", lines[1]);
+	TEST_ASSERT_EQUAL_STRING("Line 3", lines[2]);
+}
+
+void test_splitTextLines_empty_string(void) {
+	char str[] = "";
+	char* lines[MAX_TEXT_LINES];
+	int count = splitTextLines(str, lines, MAX_TEXT_LINES);
+	TEST_ASSERT_EQUAL(1, count);
+}
+
+void test_splitTextLines_null_string(void) {
+	char* lines[MAX_TEXT_LINES];
+	int count = splitTextLines(NULL, lines, MAX_TEXT_LINES);
+	TEST_ASSERT_EQUAL(0, count);
+}
+
+void test_splitTextLines_max_lines(void) {
+	char str[] = "1\n2\n3\n4\n5";
+	char* lines[MAX_TEXT_LINES];
+	int count = splitTextLines(str, lines, 3);
+	TEST_ASSERT_EQUAL(3, count); // Should stop at max_lines
+}
+
+// ============================================================================
+// Main Test Runner
+// ============================================================================
+
+int main(void) {
+	UNITY_BEGIN();
+
+	// String matching
+	RUN_TEST(test_prefixMatch_exact);
+	RUN_TEST(test_prefixMatch_prefix);
+	RUN_TEST(test_prefixMatch_no_match);
+	RUN_TEST(test_prefixMatch_case_insensitive);
+	RUN_TEST(test_prefixMatch_empty);
+
+	RUN_TEST(test_suffixMatch_exact);
+	RUN_TEST(test_suffixMatch_suffix);
+	RUN_TEST(test_suffixMatch_no_match);
+	RUN_TEST(test_suffixMatch_case_insensitive);
+	RUN_TEST(test_suffixMatch_extension);
+
+	RUN_TEST(test_exactMatch_same);
+	RUN_TEST(test_exactMatch_different);
+	RUN_TEST(test_exactMatch_case_sensitive);
+	RUN_TEST(test_exactMatch_null_strings);
+
+	RUN_TEST(test_containsString_found);
+	RUN_TEST(test_containsString_not_found);
+	RUN_TEST(test_containsString_case_insensitive);
+
+	RUN_TEST(test_hide_hidden_file);
+	RUN_TEST(test_hide_disabled_file);
+	RUN_TEST(test_hide_map_txt);
+	RUN_TEST(test_hide_normal_file);
+
+	// String manipulation
+	RUN_TEST(test_normalizeNewline_windows);
+	RUN_TEST(test_normalizeNewline_unix);
+	RUN_TEST(test_normalizeNewline_no_newline);
+
+	RUN_TEST(test_trimTrailingNewlines_single);
+	RUN_TEST(test_trimTrailingNewlines_multiple);
+	RUN_TEST(test_trimTrailingNewlines_none);
+
+	RUN_TEST(test_trimSortingMeta_with_number);
+	RUN_TEST(test_trimSortingMeta_no_number);
+	RUN_TEST(test_trimSortingMeta_with_space);
+
+	// Text line splitting
+	RUN_TEST(test_splitTextLines_single_line);
+	RUN_TEST(test_splitTextLines_multiple_lines);
+	RUN_TEST(test_splitTextLines_empty_string);
+	RUN_TEST(test_splitTextLines_null_string);
+	RUN_TEST(test_splitTextLines_max_lines);
+
+	return UNITY_END();
+}
